修复bug9206消息，alarmflag从32改为64位

+++ src/JT808.Protocol.Extensions/JT808.Protocol.Extensions.JT1078/MessageBody/JT808_0x9206.cs
@@ -58,17 +58,17 @@ namespace JT808.Protocol.Extensions.JT1078.MessageBody
         /// </summary>
         public byte ChannelNo { get; set; }
         /// <summary>
-        /// 起始时间
+        /// 起始时间 BCD[6]
         /// </summary>
         public DateTime BeginTime { get; set; }
         /// <summary>
-        /// 结束时间
+        /// 结束时间 BCD[6]
         /// </summary>
         public DateTime EndTime { get; set; }
         /// <summary>
         /// 报警标志
         /// </summary>
-        public uint AlarmFlag { get; set; }
+        public UInt64 AlarmFlag { get; set; }
         /// <summary>
         /// 音视频资源类型
         /// </summary>
@@ -119,7 +119,7 @@ namespace JT808.Protocol.Extensions.JT1078.MessageBody
             writer.WriteString($"[{value.BeginTime.ToString("yyMMddHHmmss")}]起始时间", value.BeginTime.ToString("yyyy-MM-dd HH:mm:ss"));
             value.EndTime = reader.ReadDateTime6();
             writer.WriteString($"[{value.EndTime.ToString("yyMMddHHmmss")}]起始时间", value.EndTime.ToString("yyyy-MM-dd HH:mm:ss"));
-            value.AlarmFlag = reader.ReadUInt32();
+            value.AlarmFlag = reader.ReadUInt64();
             writer.WriteNumber($"[{value.AlarmFlag.ReadNumber()}]报警标志", value.AlarmFlag);
             value.MediaType = reader.ReadByte();
             writer.WriteString($"[{value.MediaType.ReadNumber()}]音视频类型", AVResourceTypeDisplay(value.MediaType));
@@ -207,7 +207,7 @@ namespace JT808.Protocol.Extensions.JT1078.MessageBody
             jT808_0x9206.ChannelNo = reader.ReadByte();
             jT808_0x9206.BeginTime = reader.ReadDateTime6();
             jT808_0x9206.EndTime = reader.ReadDateTime6();
-            jT808_0x9206.AlarmFlag = reader.ReadUInt32();
+            jT808_0x9206.AlarmFlag = reader.ReadUInt64();
             jT808_0x9206.MediaType = reader.ReadByte();
             jT808_0x9206.StreamType = reader.ReadByte();
             jT808_0x9206.MemoryPositon = reader.ReadByte();
@@ -233,7 +233,7 @@ namespace JT808.Protocol.Extensions.JT1078.MessageBody
             writer.WriteByte(value.ChannelNo);
             writer.WriteDateTime6(value.BeginTime);
             writer.WriteDateTime6(value.EndTime);
-            writer.WriteUInt32(value.AlarmFlag);
+            writer.WriteUInt64(value.AlarmFlag);
             writer.WriteByte(value.MediaType);
             writer.WriteByte(value.StreamType);
             writer.WriteByte(value.MemoryPositon);
